Move parser to internals
